Put -Wall in CPP_FLAGS in makefile. I then fixed any errors that revealed during compilation. I removed the OverlayNewTexture member function from SnazzCraft::Texture as it is not needed now or in the percievable future.

+++ SnazzCraft/src/engine/world/world.cpp
@@ -48,7 +48,7 @@ void SnazzCraft::World::RenderChunks(SnazzCraft::User* Player)
 
     for (int32_t X = PlayerChunkPosition[0] - static_cast<int32_t>(this->RenderDistance); X <= PlayerChunkPosition[0] + static_cast<int32_t>(this->RenderDistance); X++) {
     for (int32_t Z = PlayerChunkPosition[1] - static_cast<int32_t>(this->RenderDistance); Z <= PlayerChunkPosition[1] + static_cast<int32_t>(this->RenderDistance); Z++) {
-        if (X < 0 || X >= this->Size || Z < 0 || Z >= this->Size) continue;
+        if (X < 0 || X >= static_cast<int32_t>(this->Size) || Z < 0 || Z >= static_cast<int32_t>(this->Size)) continue;
 
         auto ChunkIterator = this->Chunks->find(INDEX_2D(X, Z, static_cast<int32_t>(this->Size)));
         if (ChunkIterator == this->Chunks->end()) continue;
